Updated where PullRequestReviewFactory exists, updated some factory fields
