+++ backend/git_real/tests/factories.py
@@ -1,5 +1,5 @@
 from factory.declarations import LazyAttribute
-from git_real.models import PullRequest
+from git_real.models import PullRequest, PullRequestReview
 from core import models
 from factory.django import DjangoModelFactory
 import factory
@@ -23,9 +23,25 @@ def _number_generator():
         i += 1
 
 
+def _html_url_generator():
+    i = 1
+    while True:
+        yield f"https://somewhere_html.com/like_this{i}"
+        i += 1
+
+
+def _commit_id_generator():
+    i = 1
+    while True:
+        yield f"{i}21747298a3790fde1710f3aa2d03b55020575aa"
+        i += 1
+
+
 _repo_full_name_iterator = _repo_full_name_generator()
 _repo_ssh_url_iterator = _repo_full_name_iterator
 _number_iterator = _number_generator()
+_html_url_generator = _html_url_generator()
+_commit_id_generator = _commit_id_generator()
 
 
 class RepositoryFactory(DjangoModelFactory):
@@ -60,3 +76,16 @@ class PullRequestFactory(DjangoModelFactory):
     title = "title"
     body = "body"
     state = PullRequest.OPEN
+
+
+class PullRequestReviewFactory(DjangoModelFactory):
+    class Meta:
+        model = PullRequestReview
+
+    html_url = factory.lazy_attribute(lambda *args, **kwargs: next(_html_url_generator))
+    pull_request = factory.SubFactory(PullRequestFactory)
+    author_github_name = factory.Faker("first_name")
+    submitted_at = factory.lazy_attribute(lambda o: timezone.now())
+    body = "A horse walks into a bar, the barman says 'Why the long face?'"
+    state = "closed"
+    commit_id = factory.LazyAttribute(lambda *args, **kwargs: next(_commit_id_generator))
